Composition alternative:
If you consider Automatono to be the base prototype then it becomes
easier to create an adapter. Eventually we would be able to refactor any
base behaviors found in Probot into Automatono and any Slack connection
logic to the adapter that extends it. This would completely decouple the
adadpter from the conversation logic.

The updates that are not related to composition config extension and
mostly string interpolation or general style.